split config subcommands into files
