+++ pkg/cmd/config/config_suite_test.go
@@ -1,5 +1,6 @@
 /*
 SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
+
 SPDX-License-Identifier: Apache-2.0
 */
