+++ pkg/cmd/config/set_garden.go
@@ -20,13 +20,38 @@ import (
 	"github.com/gardener/gardenctl-v2/pkg/config"
 )
 
-// Options is a struct to support all config subcommands
-type options struct {
+// NewCmdConfigSetGarden returns a new (config) set-garden command.
+func NewCmdConfigSetGarden(f util.Factory, ioStreams util.IOStreams) *cobra.Command {
+	o := &setGardenOptions{
+		Options: base.Options{
+			IOStreams: ioStreams,
+		},
+	}
+	cmd := &cobra.Command{
+		Use:   "set-garden",
+		Short: "modify or add Garden to gardenctl configuration",
+		Long:  "Modify or add Garden to gardenctl configuration. E.g. \"gardenctl config set-garden my-garden --kubeconfig ~/.kube/kubeconfig.yaml\" to configure or add a garden with identity 'my-garden'",
+		ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+			suggestions, err := validGardenArgsFunction(f, args)
+			if err != nil {
+				fmt.Fprintln(o.IOStreams.ErrOut, err.Error())
+				return nil, cobra.ShellCompDirectiveNoFileComp
+			}
+
+			return util.FilterStringsByPrefix(toComplete, suggestions), cobra.ShellCompDirectiveNoFileComp
+		},
+		RunE: base.WrapRunE(o, f),
+	}
+
+	o.AddFlags(cmd.Flags())
+
+	return cmd
+}
+
+type setGardenOptions struct {
 	base.Options
 	// Configuration is the gardenctl configuration
 	Configuration *config.Config
-	// Command is the name of the config command
-	Command string
 	// Name is a unique name of this Garden that can be used to target this Garden
 	Name string
 	// KubeconfigFlag is the path to the kubeconfig file of the Garden cluster
@@ -42,76 +67,49 @@ type options struct {
 }
 
 // Complete adapts from the command line args to the data required.
-func (o *options) Complete(f util.Factory, cmd *cobra.Command, args []string) error {
-	config, err := getConfiguration(f)
+func (o *setGardenOptions) Complete(f util.Factory, cmd *cobra.Command, args []string) error {
+	manager, err := f.Manager()
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get target manager: %w", err)
+	}
+
+	config := manager.Configuration()
+	if config == nil {
+		return errors.New("failed to get configuration")
 	}
 
 	o.Configuration = config
 
-	switch o.Command {
-	case "set-garden", "delete-garden":
-		if len(args) > 0 {
-			o.Name = strings.TrimSpace(args[0])
-		}
-	case "view":
-		if o.Output == "" {
-			o.Output = "yaml"
-		}
+	if len(args) > 0 {
+		o.Name = strings.TrimSpace(args[0])
 	}
 
 	return nil
 }
 
 // Validate validates the provided options
-func (o *options) Validate() error {
-	switch o.Command {
-	case "set-garden", "delete-garden":
-		if o.Name == "" {
-			return errors.New("garden identity is required")
-		}
-	case "view":
-		return o.Options.Validate()
+func (o *setGardenOptions) Validate() error {
+	if o.Name == "" {
+		return errors.New("garden identity is required")
 	}
 
 	return nil
 }
 
 // AddFlags adds flags to adjust the output to a cobra command
-func (o *options) AddFlags(flags *pflag.FlagSet) {
-	switch o.Command {
-	case "set-garden":
-		flags.Var(&o.KubeconfigFlag, "kubeconfig", "path to kubeconfig file for this Garden cluster")
-		flags.Var(&o.ContextFlag, "context", "override the current-context of the garden cluster kubeconfig")
-		flags.StringArrayVar(&o.Patterns, "pattern", nil, `define regex match patterns for this garden for custom input formats for targeting.
+func (o *setGardenOptions) AddFlags(flags *pflag.FlagSet) {
+	flags.Var(&o.KubeconfigFlag, "kubeconfig", "path to kubeconfig file for this Garden cluster")
+	flags.Var(&o.ContextFlag, "context", "override the current-context of the garden cluster kubeconfig")
+	flags.StringArrayVar(&o.Patterns, "pattern", nil, `define regex match patterns for this garden for custom input formats for targeting.
 Use named capturing groups to match target values.
 Supported capturing groups: project, namespace, shoot.
 Note that if you set this flag it will overwrite the pattern list in the config file.
 You may specify any number of extra patterns.
 Example: ^((?Pmy-garden[^/]+)/)?shoot--(?P<project>.+)--(?P<shoot>.+)$`)
-	case "delete-garden":
-		// no flags
-	case "view":
-		o.Options.AddFlags(flags)
-	}
 }
 
 // Run executes the command
-func (o *options) Run(_ util.Factory) error {
-	switch o.Command {
-	case "set-garden":
-		return o.runSetGarden()
-	case "delete-garden":
-		return o.runDeleteGarden()
-	case "view":
-		return o.runView()
-	}
-
-	return nil
-}
-
-func (o *options) runSetGarden() error {
+func (o *setGardenOptions) Run(_ util.Factory) error {
 	garden, err := o.Configuration.Garden(o.Name)
 	if err == nil {
 		if o.KubeconfigFlag.Provided() {
@@ -147,24 +145,4 @@ func (o *options) runSetGarden() error {
 	fmt.Fprintf(o.IOStreams.Out, "Successfully configured garden %q\n", o.Name)
 
 	return nil
-}
-
-func (o *options) runDeleteGarden() error {
-	err := o.Configuration.DeleteGarden(o.Name)
-	if err != nil {
-		return err
-	}
-
-	err = o.Configuration.Save()
-	if err != nil {
-		return fmt.Errorf("failed to delete garden from configuration: %w", err)
-	}
-
-	fmt.Fprintf(o.IOStreams.Out, "Successfully deleted garden %q\n", o.Name)
-
-	return nil
-}
-
-func (o *options) runView() error {
-	return o.PrintObject(o.Configuration)
 }
